requirements.txt
# GEO pipeline dependencies
python-dotenv>=1.0
langgraph>=0.2
groq>=0.9
mistralai>=1.0
pandas>=2.0
tabulate>=0.9
rapidfuzz>=3.0
httpx>=0.27
requests>=2.31
ddgs>=6.0
wikipedia-api>=0.6
apify-client>=1.7
google-search-results>=2.4   # SerpApi
